Automatically upload checksumed taball on release
This commit instructs Travis to upload a compressed tarball
with the same contents (essentially) as GitHub's compressed tarball.
It also uploads an md5sum for the generated compressed tarball as well.

The upload happens on tags, when the Python version is Python 3.4 and
the krb5 version is 1.13 (this is intended to stay as the latest
published release).  To facilitate this, the KRB5_VER environment
variable was added to each environment in the test matrix.

To ensure that cleanup happens properly, a couple of extra directories
were added to the .gitignore file.